@@ -2,7 +2,7 @@
 //  BarStyleView.swift
 //  Example (iOS)
 //
-//  Copyright © 2021 Xmartlabs SRL. All rights reserved.
+//  Copyright © 2022 Xmartlabs SRL. All rights reserved.
 //
 
 import SwiftUI
@@ -11,22 +11,27 @@ import PagerTabStripView
 struct BarStyleView: View {
     @State var selection = 1
 
-    @ObservedObject var tweetsModel = TweetsModel()
-    @ObservedObject var mediaModel = MediaModel()
-    @ObservedObject var likesModel = LikesModel()
+    @StateObject var tweetsModel = TweetsModel()
+    @StateObject var mediaModel = TweetsModel()
+    @StateObject var likesModel = TweetsModel()
 
     @MainActor var body: some View {
         PagerTabStripView(selection: $selection) {
-            PostsList(isLoading: $tweetsModel.isLoading, items: tweetsModel.posts).pagerTabItem {
-            }
+            PostsList(isLoading: $tweetsModel.isLoading, items: tweetsModel.posts)
+                .pagerTabItem(tag: 0) {
+                }
 
-            PostsList(isLoading: $mediaModel.isLoading, items: mediaModel.posts).pagerTabItem {
-            }
+            PostsList(isLoading: $mediaModel.isLoading, items: mediaModel.posts)
+                .pagerTabItem(tag: 1) {
+                }
 
-            PostsList(isLoading: $likesModel.isLoading, items: likesModel.posts, withDescription: false).pagerTabItem {
-            }
+            PostsList(isLoading: $likesModel.isLoading, items: likesModel.posts, withDescription: false)
+                .pagerTabItem(tag: 2) {
+                }
         }
-        .pagerTabStripViewStyle(.bar(placedInToolbar: false, indicatorViewHeight: 6) { Rectangle().fill(.yellow) })
+        .pagerTabStripViewStyle(.bar(placedInToolbar: false, indicatorViewHeight: 6) {
+            Rectangle().fill(.yellow)
+        })
         .navigationTitle("Bar Style View")
     }
 }
