@@ -2,7 +2,7 @@
 //  InstagramView.swift
 //  Example (iOS)
 //
-//  Copyright © 2021 Xmartlabs SRL. All rights reserved.
+//  Copyright © 2022 Xmartlabs SRL. All rights reserved.
 //
 
 import SwiftUI
@@ -10,54 +10,75 @@ import PagerTabStripView
 
 struct InstagramView: View {
 
-    @State var selection = 1
+    enum Page: String {
+        case gallery = "photo.stack"
+        case list  = ""
+        case like
+        case saved
+    }
 
-    @ObservedObject var galleryModel = GalleryModel()
-    @ObservedObject var listModel = ListModel()
-    @ObservedObject var likedModel = LikedModel()
-    @ObservedObject var savedModel = SavedModel()
+    @State var selection = Page.list
+    @State var toggle = true
+
+    @StateObject var galleryModel = ListModel()
+    @StateObject var listModel = ListModel()
+    @StateObject var likedModel = ListModel()
+    @StateObject var savedModel = ListModel()
 
     @MainActor var body: some View {
         PagerTabStripView(selection: $selection) {
-            PostsList(isLoading: $galleryModel.isLoading, items: galleryModel.posts).pagerTabItem {
-                InstagramNavBarItem(imageName: "gallery")
-            }.onPageAppear {
-                galleryModel.isLoading = true
-                DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
-                    galleryModel.isLoading = false
+            PostsList(isLoading: $galleryModel.isLoading, items: galleryModel.posts)
+                .pagerTabItem(tag: Page.gallery) {
+                    InstagramNavBarItem(imageName: "photo.stack", selection: $selection, tag: Page.gallery)
                 }
-            }
-
-            PostsList(isLoading: $listModel.isLoading, items: listModel.posts, withDescription: false).pagerTabItem {
-                InstagramNavBarItem(imageName: "list")
-            }.onPageAppear {
-                listModel.isLoading = true
-                DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
-                    listModel.isLoading = false
+            PostsList(isLoading: $listModel.isLoading, items: listModel.posts, withDescription: false)
+                .pagerTabItem(tag: Page.list) {
+                    InstagramNavBarItem(imageName: "chart.bar.doc.horizontal" /* "list.bullet" */, selection: $selection, tag: Page.list)
                 }
-            }
-
-            PostsList(isLoading: $likedModel.isLoading, items: likedModel.posts).pagerTabItem {
-                InstagramNavBarItem(imageName: "liked")
-            }.onPageAppear {
-                likedModel.isLoading = true
-                DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
-                    likedModel.isLoading = false
-                }
-            }
-
-            PostsList(isLoading: $savedModel.isLoading, items: savedModel.posts, withDescription: false).pagerTabItem {
-                InstagramNavBarItem(imageName: "saved")
-            }.onPageAppear {
-                savedModel.isLoading = true
-                DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
-                    savedModel.isLoading = false
-                }
+            if toggle {
+                PostsList(isLoading: $likedModel.isLoading, items: likedModel.posts)
+                    .pagerTabItem(tag: Page.like) {
+                        InstagramNavBarItem(imageName: "heart", selection: $selection, tag: Page.like)
+                    }
+                PostsList(isLoading: $savedModel.isLoading, items: savedModel.posts, withDescription: false)
+                    .pagerTabItem(tag: Page.saved) {
+                        InstagramNavBarItem(imageName: "photo.stack" /* "bookmark"*/, selection: $selection, tag: Page.saved)
+                    }
             }
         }
-        .pagerTabStripViewStyle(.barButton(placedInToolbar: false, pagerAnimation: .default, tabItemHeight: 50, indicatorView: {
-            Rectangle().fill(Color(.systemGray))
-        }))
+        .pagerTabStripViewStyle(.barButton(placedInToolbar: false, pagerAnimation: .default,
+                                           tabItemHeight: 50, indicatorViewHeight: 2, indicatorView: {
+                                            Rectangle().fill(.blue).cornerRadius(1)
+                                           }))
+        .navigationBarItems(trailing: Button("Refresh") {
+            toggle.toggle()
+        })
+    }
+}
+
+struct InstagramNavBarItem<SelectionType>: View where SelectionType: Hashable {
+    @EnvironmentObject private var pagerSettings: PagerSettings<SelectionType>
+
+    var image: Image
+    @Binding var selection: SelectionType
+    let tag: SelectionType
+
+    init(imageName: String, selection: Binding<SelectionType>, tag: SelectionType) {
+        self.tag = tag
+        self.image = Image(systemName: imageName)
+        _selection = selection
+    }
+
+    @MainActor var body: some View {
+        VStack {
+            image
+                .renderingMode(.template)
+                .resizable()
+                .frame(width: 25.0, height: 25)
+                .foregroundColor(.gray.interpolateTo(color: .blue, fraction: pagerSettings.transition.progress(for: tag)))
+        }
+        .animation(.easeInOut, value: selection)
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
     }
 }
 
