@@ -2,7 +2,7 @@
 //  ExampleApp.swift
 //  Shared
 //
-//  Copyright © 2021 Xmartlabs SRL. All rights reserved.
+//  Copyright © 2022 Xmartlabs SRL. All rights reserved.
 //
 
 import SwiftUI
