@@ -2,7 +2,7 @@
 //  User.swift
 //  Example (iOS)
 //
-//  Copyright © 2021 Xmartlabs SRL. All rights reserved.
+//  Copyright © 2022 Xmartlabs SRL. All rights reserved.
 //
 
 import Foundation
@@ -10,8 +10,4 @@ import Foundation
 struct User: Hashable, Codable, Identifiable {
     var id: Int
     var name: String
-    var imageURL: String
-    var image: URLImageView {
-        URLImageView(urlString: imageURL)
-    }
 }
