@@ -2,7 +2,7 @@
 //  PostDetail.swift
 //  Example (iOS)
 //
-//  Copyright © 2021 Xmartlabs SRL. All rights reserved.
+//  Copyright © 2022 Xmartlabs SRL. All rights reserved.
 //
 
 import SwiftUI
@@ -11,27 +11,26 @@ struct PostDetail: View {
     var post: Post
 
     @MainActor var body: some View {
-        ScrollView {
-            post.user.image
-                .frame(width: 100.0, height: 100.0)
-                .clipShape(Circle())
-                .overlay(Circle().stroke(Color.white, lineWidth: 4))
-                .shadow(radius: 7)
-                .padding(.top, 100)
-
+        ScrollView(.vertical, showsIndicators: true) {
             VStack(alignment: .center) {
+                Image("defaultImage")
+                    .resizable()
+                    .frame(width: 100.0, height: 100.0)
+                    .clipShape(Circle())
+                    .overlay(Circle().stroke(Color.white, lineWidth: 4))
+                    .shadow(radius: 7)
+                    .padding(.top, 100)
+                Spacer()
                 Text(post.user.name)
                     .font(.title)
                     .foregroundColor(.primary)
-
                 Spacer()
-
                 Text(post.text)
                     .font(.headline)
                     .foregroundColor(.secondary)
                     .multilineTextAlignment(.center)
             }
-            .padding()
+            .frame( maxWidth: .infinity)
         }
     }
 }
