@@ -2,7 +2,7 @@
 //  PagerTabStripView.h
 //  PagerTabStripView
 //
-//  Copyright © 2021 Xmartlabs SRL. All rights reserved.
+//  Copyright © 2022 Xmartlabs SRL. All rights reserved.
 //
 
 #import <UIKit/UIKit.h>
