@@ -2,42 +2,28 @@
 //  NavBarItem.swift
 //  PagerTabStripView
 //
-//  Copyright © 2021 Xmartlabs SRL. All rights reserved.
+//  Copyright © 2022 Xmartlabs SRL. All rights reserved.
 //
 
 import SwiftUI
 
-struct NavBarItem: View, Identifiable {
-    @EnvironmentObject private var dataStore: DataStore
-    @Binding private var selection: Int
-    internal var id: Int
+struct NavBarItem<SelectionType>: View, Identifiable where SelectionType: Hashable {
 
-    public init(id: Int, selection: Binding<Int>) {
-        self._selection = selection
+    var id: SelectionType
+    @Binding private var selection: SelectionType
+    @EnvironmentObject private var pagerSettings: PagerSettings<SelectionType>
+
+    public init(id: SelectionType, selection: Binding<SelectionType>) {
         self.id = id
+        self._selection = selection
     }
 
     @MainActor var body: some View {
-        if id < dataStore.itemsCount {
-            VStack {
-                Button(action: {
+        if let dataItem = pagerSettings.items[id] {
+            dataItem.view
+                .onTapGesture {
                     selection = id
-                }, label: {
-                    dataStore.items[id]?.view
-                })
-                .buttonStyle(.plain)
-                .onLongPressGesture(minimumDuration: 0, maximumDistance: .infinity) { pressing in
-                    dataStore.items[id]?.tabViewDelegate?.setState(state: pressing ? .highlighted : (id == selection ? .selected : .normal))
-                } perform: {}
-            }.background(
-                GeometryReader { geometry in
-                    Color.clear.onAppear {
-                        dataStore.items[id]?.itemWidth = geometry.size.width
-                        let widthUpdated = dataStore.items.filter({ $0.value.itemWidth ?? 0 > 0 }).count == dataStore.itemsCount
-                        dataStore.widthUpdated = dataStore.itemsCount > 0 && widthUpdated
-                    }
                 }
-            )
         }
     }
 }
