@@ -2,145 +2,152 @@
 //  ScrollableNavBarView.swift
 //  PagerTabStripView
 //
-//  Created by Cecilia Pirotto on 23/8/21.
+//  Copyright © 2022 Xmartlabs SRL. All rights reserved.
 //
 
 import Foundation
 import SwiftUI
 
-internal struct ScrollableNavBarView: View {
-    @Binding private var selection: Int
-    @State private var switchAppeared: Bool = false
+internal struct ScrollableNavBarView<SelectionType>: View where SelectionType: Hashable {
 
-    @EnvironmentObject private var dataStore: DataStore
+    @Binding var selection: SelectionType
+    @EnvironmentObject private var pagerSettings: PagerSettings<SelectionType>
+    @Environment(\.pagerStyle) private var style: PagerStyle
+    @State private var appeared = false
 
-    public init(selection: Binding<Int>) {
+    public init(selection: Binding<SelectionType>) {
         self._selection = selection
     }
 
     @MainActor var body: some View {
         if let internalStyle = style as? BarButtonStyle {
-            ScrollViewReader { value in
+            ScrollViewReader { proxy in
                 ScrollView(.horizontal, showsIndicators: false) {
-                    VStack {
-                        HStack(spacing: internalStyle.tabItemSpacing) {
-                            ForEach(0..<dataStore.itemsCount, id: \.self) { idx in
-                                NavBarItem(id: idx, selection: $selection)
-                            }
+                    ScrollableNavBarViewLayout(spacing: internalStyle.tabItemSpacing) {
+                        ForEach(pagerSettings.itemsOrderedByIndex, id: \.self) { tag in
+                            NavBarItem(id: tag, selection: $selection)
+                                .tag(tag)
                         }
-                        IndicatorScrollableBarView(selection: $selection)
+                        internalStyle.indicatorView()
+                            .frame(height: internalStyle.indicatorViewHeight)
+                            .layoutValue(key: PagerWidth.self, value: pagerSettings.width)
+                            .layoutValue(key: PagerOffset.self, value: pagerSettings.contentOffset)
+                            .animation(appeared ? .default : .none, value: pagerSettings.contentOffset)
                     }
                     .frame(height: internalStyle.tabItemHeight)
                 }
+                .background(internalStyle.barBackgroundView())
                 .padding(internalStyle.padding)
-                .onChange(of: switchAppeared) { _ in
-                    // This is necessary because anchor: .center is not working correctly
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                        var remainingItemsWidth = dataStore.items[selection]?.itemWidth ?? 0 / 2
-                        let items = dataStore.items.filter { index, _ in
-                            index > selection
-                        }
-                        remainingItemsWidth += items.map {$0.value.itemWidth ?? 0}.reduce(0, +)
-                        remainingItemsWidth += CGFloat(dataStore.items.count-1 - selection)*internalStyle.tabItemSpacing
-                        let centerSel = remainingItemsWidth > settings.width/2
-                        value.scrollTo(centerSel ? selection : dataStore.items.count-1, anchor: centerSel ? .center : nil)
+                .onChange(of: pagerSettings.itemsOrderedByIndex) { _ in
+                    if pagerSettings.items[selection] != nil {
+                        proxy.scrollTo(selection, anchor: .center)
                     }
                 }
                 .onChange(of: selection) { newSelection in
                     withAnimation {
-                        value.scrollTo(newSelection, anchor: .center)
+                        if pagerSettings.items[newSelection] != nil {
+                            proxy.scrollTo(newSelection, anchor: .center)
+                        }
                     }
                 }
-            }
-            .onAppear {
-                switchAppeared = !switchAppeared
+                .onAppear {
+                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
+                        if pagerSettings.items[selection] != nil {
+                            proxy.scrollTo(selection, anchor: .center)
+                        }
+                        appeared = true
+                    }
+                }
+                .onDisappear {
+                    appeared = false
+                }
             }
         }
     }
-
-    @Environment(\.pagerStyle) var style: PagerStyle
-    @EnvironmentObject private var settings: PagerSettings
 }
 
-internal struct IndicatorScrollableBarView: View {
-    @EnvironmentObject private var dataStore: DataStore
-    @Binding private var selection: Int
-    @State private var position: Double = 0
-    @State private var selectedItemWidth: Double = 0
-    @State private var appeared: Bool = false
+struct ScrollableNavBarViewLayout: Layout {
 
-    public init(selection: Binding<Int>) {
-        self._selection = selection
+    private let spacing: CGFloat
+
+    init(spacing: CGFloat) {
+        self.spacing = spacing
     }
 
-    @MainActor var body: some View {
-        if let internalStyle = style as? BarButtonStyle {
-            internalStyle.indicatorView()
-                .frame(height: internalStyle.indicatorViewHeight)
-                .animation(.none, value: appeared)
-                .frame(width: selectedItemWidth)
-                .position(x: position)
-                .onAppear {
-                    appeared = true
-                }
-                .onChange(of: dataStore.widthUpdated) { updated in
-                    if updated {
-                        let items = dataStore.items.filter { index, _ in
-                            index < selection
-                        }
-                        selectedItemWidth = dataStore.items[selection]?.itemWidth ?? 0
-                        var newPosition = items.map({return $0.value.itemWidth ?? 0}).reduce(0, +)
-                        newPosition += (internalStyle.tabItemSpacing * CGFloat(selection)) + selectedItemWidth/2
-                        position = newPosition
-                    }
-                }
-                .onChange(of: settings.contentOffset) { newValue in
-                    let offset = newValue + (settings.width * CGFloat(selection))
-                    let percentage = offset / settings.width
-                    let items = dataStore.items.filter { index, _ in
-                        index < selection
-                    }
+    func sizeThatFits(
+        proposal: ProposedViewSize,
+        subviews: Subviews,
+        cache: inout ()
+    ) -> CGSize {
+        // Calculate and return the size of the layout container.
+        var tabBarIndices = subviews.indices
+        let indicatorIndex = tabBarIndices.removeLast()
+        let tabBarViews = subviews[tabBarIndices]
+        let indicatorSubview = subviews[indicatorIndex]
 
-                    let spaces = internalStyle.tabItemSpacing * CGFloat(selection-1)
-                    let actualWidth = dataStore.items[selection]?.itemWidth ?? 0
-                    var lastPosition = items.map({return $0.value.itemWidth ?? 0}).reduce(0, +)
-                    lastPosition += spaces + actualWidth/2
-                    var nextPosition = items.map({return $0.value.itemWidth ?? 0}).reduce(0, +)
-                    if percentage == 0 {
-                        selectedItemWidth = dataStore.items[selection]?.itemWidth ?? 0
-                        var newPosition = items.map({return $0.value.itemWidth ?? 0}).reduce(0, +)
-                        newPosition += internalStyle.tabItemSpacing * CGFloat(selection) + selectedItemWidth/2
-                        position = newPosition
-                    } else {
-                        if percentage < 0 {
-                            nextPosition += actualWidth + internalStyle.tabItemSpacing * CGFloat(selection+1)
-                            nextPosition += ((dataStore.items[selection + 1])?.itemWidth ?? 0)/2
-                        } else {
-                            nextPosition += internalStyle.tabItemSpacing * CGFloat(selection-1)
-                            nextPosition -= ((dataStore.items[selection - 1])?.itemWidth ?? 0)/2
-                        }
-                        position = lastPosition + (nextPosition - lastPosition)*abs(percentage)
+        let subviewSize = tabBarViews.map { $0.sizeThatFits(.unspecified) }
+        let maxHeight = subviewSize.map { $0.height }.reduce(CGFloat.zero) { max($0, $1) }
+        let fullSpacing = tabBarViews.count > 1 ?  CGFloat(tabBarViews.count - 1) * self.spacing : CGFloat.zero
+        let height = proposal.replacingUnspecifiedDimensions().height
+        return CGSize(width: subviewSize.map { $0.width }.reduce(0, +) + fullSpacing,
+                      height: max(height, maxHeight + indicatorSubview.sizeThatFits(.unspecified).height))
+    }
 
-                        if let selectedWidth = dataStore.items[selection]?.itemWidth,
-                           let nextWidth = percentage > 0 ? dataStore.items[selection-1]?.itemWidth : dataStore.items[selection+1]?.itemWidth,
-                           abs(percentage)>0 {
-                            selectedItemWidth = selectedWidth - (selectedWidth-nextWidth)*abs(percentage)
-                        }
-                    }
+    func placeSubviews(in bounds: CGRect, proposal: ProposedViewSize, subviews: Subviews, cache: inout ()) {
+        var tabBarIndices = subviews.indices
+        let indicatorIndex = tabBarIndices.removeLast()
+        let tabBarViews = subviews[tabBarIndices]
+        let indicatorSubview = subviews[indicatorIndex]
+        let indicatorViewSize =  indicatorSubview.sizeThatFits(.unspecified)
 
-                }
-                .onChange(of: selection) { newValue in
-                    let items = dataStore.items.filter { index, _ in
-                        index < newValue
-                    }
-                    selectedItemWidth = dataStore.items[newValue]?.itemWidth ?? 0
-                    var newPosition = items.map({return $0.value.itemWidth ?? 0}).reduce(0, +)
-                    newPosition += (internalStyle.tabItemSpacing * CGFloat(newValue)) + selectedItemWidth/2
-                    position = newPosition
-                }
+        if tabBarViews.count == 0 { return }
+        let subviewSize = tabBarViews.map { $0.sizeThatFits(.unspecified) }
+        let maxHeight = subviewSize.map { $0.height }.reduce(CGFloat.zero, max)
+        var x = bounds.minX
+        var itemsFrames = [CGRect]()
+
+        for index in tabBarIndices.indices {
+            let spacing = index < tabBarIndices.indices.last! ? self.spacing : CGFloat.zero
+            let subview = subviews[index]
+            x += subviewSize[index].width / 2
+            let point = CGPoint(x: x, y: bounds.midY - (indicatorViewSize.height / 2))
+            let proposedSize = ProposedViewSize(width: subviewSize[index].width, height: maxHeight)
+            subview.place(at: point, anchor: .center, proposal: proposedSize)
+            itemsFrames.append(CGRect(origin: point, size: CGSize(width: proposedSize.width!, height: proposedSize.height!)))
+            x += subviewSize[index].width / 2 + spacing
         }
+
+        let contentOffset = -indicatorSubview[PagerOffset.self]
+        let itemsCount = tabBarViews.count
+        let pagerWidth = indicatorSubview[PagerWidth.self]
+        let indexAndPercentage = contentOffset / pagerWidth
+        let percentage = (indexAndPercentage + 1).truncatingRemainder(dividingBy: 1)
+        let lowIndex = floor(indexAndPercentage)
+        guard lowIndex < CGFloat(itemsCount) else {
+            indicatorSubview.place(at: CGPoint(x: 0, y: bounds.maxY - (indicatorViewSize.height / 2)),
+                                   anchor: .center,
+                                   proposal: ProposedViewSize.zero)
+            return
+        }
+        let currentWidth = itemsFrames[max(0, Int(lowIndex))].size.width
+        let nextWidth = itemsFrames[min(itemsCount - 1, Int(lowIndex + 1))].size.width
+        let currentPosition = lowIndex >= 0 ? itemsFrames[Int(lowIndex)].origin.x : itemsFrames[0].origin.x - currentWidth
+        let nextPosition = lowIndex + 1 < CGFloat(itemsCount - 1)
+            ? itemsFrames[Int(lowIndex + 1)].origin.x
+            : itemsFrames[Int(itemsCount - 1)].origin.x + nextWidth
+        let proposedWidth = currentWidth + ((nextWidth - currentWidth) * percentage)
+        let proposedPosition = currentPosition + ((nextPosition - currentPosition) * percentage)
+        indicatorSubview.place(at: CGPoint(x: proposedPosition, y: bounds.maxY - (indicatorViewSize.height / 2)),
+                               anchor: .center,
+                               proposal: ProposedViewSize(width: proposedWidth, height: indicatorViewSize.height))
     }
 
-    @Environment(\.pagerStyle) var style: PagerStyle
-    @EnvironmentObject private var settings: PagerSettings
+}
+
+struct PagerOffset: LayoutValueKey {
+    static let defaultValue: CGFloat = 0
+}
+
+struct PagerWidth: LayoutValueKey {
+    static let defaultValue: CGFloat = 0
 }
